Renamne format_HEMCO.py to format_hemco_data.py + other updates
gcpy/__init__.py
- Added "from .format_hemco_data" import *" to make sure that
  functions are imported from format_hemco_data.py

gcpy/format_HEMCO.py
- Renamed to format_hemco_data.py

gcpy/format_hemco_data.py
- Added code Y formatting updates suggested by Pylint
- Added _update_variable_attributes function so that we replace
  or update variable attributes to COARDS-conforming values
  without clobbering other attributes that may be present
- Updated imports statements to the proper order (suggested by Pylint)
- Added PyDoc header at the top of the module
- Renamed "t" variable to "term" to be snake_case conforming

CHANGELOG.md
- updated accordingly
