+++ gcpy/__init__.py
@@ -27,3 +27,4 @@ from .benchmark import *
 from .file_regrid import *
 from .grid_stretching_transforms import *
 from .cstools import *
+from .format_hemco_data import *
